feat: add comment custom hook in feature
